InitFineMaps    removed testing print statements channel_grid:
These modifications were made because the aml script used to create
the stream map file didn't always put channel intersections in the
same grid cell, if they were close to a grid cell boundary.

three subroutines were modified:
1. RouteDebrisFlow
2. Find_Next_Segment
3. channel_grid_has_intersection

The variable, Flag, in RouteDebrisFlow is initially set to 0.
The first thing this subroutine does is try to find an intersection of
current segment and next segment by assuming that the .aml script did
everything correctly and found the intersections for stream_map.dat.
If the search radius exceeds MaxRadius (set to 200 here, max number
of cells per segment in the Little Wenatchee is 157) then RouteDebrisFlow sets
Flag to 2, and tries to find the intersection by an alternate method.  First,
it finds the col (next_inti) and row (next_intj) for the closest cell in the
next segment. Once this is found, Flag is set to 3, and finds
the the closest cell (curr_inti,curr_intj) to next_inti and next_intj that is
in the current segment.  These four integers are then passed to
Find_Next_Segment in which aspect is looked at subroutine decides whether to
dump the sediment in the current segment or pass it onto the next.

 ----------------------------------------------------------------------