These changes were made to sedoutflow, which is used to determine erosion from pixels with channels.
A check on the available water was added. The amount of the available water is
different. It is now a function of Runon and IExcessSed.

Depth for routing is IExcessSed and not IExcessSed+IExcess.

Also, sedoutflow is used to determine Runoff. Therefore Runoff is no longer
truelly Runoff. Instead, is the calculated Runoff that is used to determine
the stable subtimestep.

 ----------------------------------------------------------------------